Reorganized the project and updated island model.
